+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 0.6.7+3
+## 0.6.8+1
 
 * Re-lands support for Impeller.
+
+## 0.6.8
+
+* Updates Guava version to 33.3.0.
 
 ## 0.6.7+2